create GrantsTableWrapper for smaller paginated fetches
